requirements.txt
pytest>=7.2.0
selene>=2.0.0b16
selenium~=4.4.3
allure-pytest~=2.12.0